switch to new API lib
