+++ pom.xml
@@ -33,7 +33,7 @@
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
         <java.version>21</java.version>
         <openapi-version>2.8.4</openapi-version>
-        <aicockpit-api>0.0.9</aicockpit-api>
+        <aicockpit-api>0.1.0</aicockpit-api>
     </properties>
 
     <dependencies>
