+++ src/main/java/de/starwit/services/ModuleDataService.java
@@ -106,79 +106,6 @@ public class ModuleDataService {
         }
     }
 
-    public ValidationFeedback validateModuleData(Module module) {
-        ValidationFeedback feedback = new ValidationFeedback();
-        feedback.setValid(true);
-
-        for (String sbom : module.getsBOMLocation().keySet()) {
-            try {
-                if(module.getsBOMLocation().get(sbom) == null) {
-                    feedback.getInvalidUris().add(sbom);
-                    feedback.setValid(false);
-                    continue;
-                }
-                URL sbomLocation = new URI(module.getsBOMLocation().get(sbom).getUrl()).toURL();
-                if( !checksURLAvailability(sbomLocation)) {
-                    feedback.getUnreachableUris().add(sbomLocation);
-                    feedback.setValid(false);
-                }
-            } catch (URISyntaxException | MalformedURLException | IllegalArgumentException e) {
-                log.debug("Invalid URI for sbom {} of module {}", sbom, module.getName());
-                feedback.getInvalidUris().add(sbom + ": " + module.getsBOMLocation().get(sbom).getUrl());
-                feedback.setValid(false);
-            }
-        }
-
-        if(module.getUseAI()) {
-            feedback.setHasIncompleteModelData(false);
-            if(module.getModel() == null || module.getModel().getModelLink() == null) {
-                log.debug("No model data for module {}", module.getName());
-                feedback.setValid(false);
-            } 
-
-            if(module.getModel().getModelLink() == null) {
-                log.debug("No model link for module {}", module.getName());
-                feedback.setHasIncompleteModelData(true);
-            } else {
-                try {
-                    URL modelLink = module.getModel().getModelLink().toURL();
-                    if(!checksURLAvailability(modelLink)) {
-                        feedback.getUnreachableUris().add(modelLink);
-                        feedback.setHasIncompleteModelData(true);
-                    }
-                    URL trainingDataLink = module.getModel().getModelLink().toURL();
-                    if(module.getModel().getPublicTrainingData() && !checksURLAvailability(trainingDataLink)) {
-                        feedback.getUnreachableUris().add(trainingDataLink);
-                        feedback.setHasIncompleteModelData(true);
-                    }
-                } catch (MalformedURLException e) {
-                    log.info("Invalid model link for module {}", module.getName());
-                    feedback.setHasIncompleteModelData(true);
-                    feedback.getInvalidUris().add(module.getModel().getModelLink().toString());
-                }              
-            }
-        }
-
-        //check if successor/submodule relations work
-        return feedback;
-    }
-
-    private boolean checksURLAvailability(URL url) {
-        try {
-            HttpURLConnection huc = (HttpURLConnection) url.openConnection();
-            if(huc.getResponseCode() == HttpURLConnection.HTTP_OK) {
-                return true;
-            }
-        } catch (IOException e) {
-            log.debug("Could not connect to {}", url.toString());
-            return false;
-        }
-        return false;
-    }
-
-    public List<Module> getModules() {
-        return modules;
-    } 
 
     public Module findModuleByName(String name) {
         for (Module module : modules) {
